pytest, pydantic 追加
